backend/requirements.txt
strawberry-graphql==0.217.1
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
pymongo==4.7.1
motor==3.3.2